+++ src/mygrad/tensor_manip/tensor_joining/funcs.py
@@ -5,9 +5,9 @@ from numpy import ndarray
 from mygrad.tensor_base import Tensor, implements_numpy_override
 from mygrad.typing import ArrayLike, DTypeLikeReals
 
-from .ops import Concatenate
+from .ops import Concatenate, Stack
 
-__all__ = ["concatenate"]
+__all__ = ["concatenate", "stack"]
 
 
 @implements_numpy_override
@@ -16,10 +16,11 @@ def concatenate(
     axis: Optional[int] = 0,
     out: Optional[Union[ndarray, Tensor]] = None,
     dtype: Optional[DTypeLikeReals] = None,
+    *,
     constant: Optional[bool] = None,
 ) -> Tensor:
     """
-    concatenate((t1, t2, ...), axis=0, out=None)
+    concatenate((t1, t2, ...), axis=0, out=None, *, constant=None)
 
     Join a sequence of tensors along an existing axis.
 
@@ -63,10 +64,10 @@ def concatenate(
 
     See Also
     --------
-    stack : Stack a sequence of arrays along a new axis.
-    hstack : Stack arrays in sequence horizontally (column wise).
-    vstack : Stack arrays in sequence vertically (row wise).
-    dstack : Stack arrays in sequence depth wise (along third dimension).
+    stack : Stack a sequence of tensors along a new axis.
+    hstack : Stack tensors in sequence horizontally (column wise).
+    vstack : Stack tensors in sequence vertically (row wise).
+    dstack : Stack tensors in sequence depth wise (along third dimension).
 
     References
     ----------
@@ -91,6 +92,83 @@ def concatenate(
         Concatenate,
         *tensors,
         op_kwargs={"axis": axis, "dtype": dtype},
+        constant=constant,
+        out=out,
+    )
+
+
+@implements_numpy_override
+def stack(
+    tensors: Sequence[ArrayLike],
+    axis: int = 0,
+    out: Optional[Union[ndarray, Tensor]] = None,
+    *,
+    constant: Optional[bool] = None,
+) -> Tensor:
+    """
+    stack((t1, t2, ...), axis=0, out=None, *, constant=None)
+
+    Join a sequence of tensors along a new axis.
+
+    This docstring was adapted from that of numpy.stack [1]_
+
+    Parameters
+    ----------
+    tensors : Sequence[ArrayLike]
+        Each tensor must have the same shape.
+
+    axis : Optional[int]
+        The axis in the result tensor along which the input tensors are stacked.
+
+    out : Optional[Union[ndarray, Tensor]]
+        If provided, the destination to place the result. The shape must be
+        correct, matching that of what concatenate would have returned if no
+        out argument were specified.
+
+    constant : Optional[bool]
+        If ``True``, this tensor is treated as a constant, and thus does not
+        facilitate back propagation (i.e. ``constant.grad`` will always return
+        ``None``).
+
+        Defaults to ``False`` for float-type data.
+        Defaults to ``True`` for integer-type data.
+
+        Integer-type tensors must be constant.
+
+    Returns
+    -------
+    res : Tensor
+        The stacked tensor has one more dimension than the input arrays.
+
+    See Also
+    --------
+    concatenate : Join a sequence of tensors along an existing axis.
+    hstack : Stack tensors in sequence horizontally (column wise).
+    vstack : Stack tensors in sequence vertically (row wise).
+    dstack : Stack tensors in sequence depth wise (along third dimension).
+
+    References
+    ----------
+    .. [1] Retrieved from https://numpy.org/doc/stable/reference/generated/numpy.stack.html
+
+    Examples
+    --------
+    >>> import mygrad as mg
+    >>> a = mg.tensor([1, 2, 3])
+    >>> b = mg.tensor([-1, -2, -3])
+    >>> mg.stack((a, b))
+    Tensor([[ 1,  2,  3],
+            [-1, -2, -3]])
+
+    >>> mg.stack((a, b), axis=-1)
+    Tensor([[1, -1],
+            [2, -2],
+            [3, -3]])
+    """
+    return Tensor._op(
+        Stack,
+        *tensors,
+        op_kwargs={"axis": axis},
         constant=constant,
         out=out,
     )